pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nautobot-topology"
version = "1.0.0"
description = "Nautobot Topology Visualization Plugin"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Joakim Nyden", email = "joakim@nydenconsulting.se" },
]
license = { text = "Apache-2.0" }
keywords = ["nautobot", "nautobot-plugin", "topology", "visualization", "network"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Networking",
]
dependencies = [
    "nautobot>=3.0.8",
]

[project.optional-dependencies]
dev = [
    "invoke",
    "black",
    "flake8",
    "pylint",
]
test = [
    "pytest",
    "pytest-django",
]

[project.entry-points."nautobot_plugins"]
nautobot_topology = "nautobot_topology:config"

[tool.hatch.build.targets.wheel]
packages = ["nautobot_topology"]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
django_settings_module = "nautobot.core.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v"

[tool.pylint.master]
load-plugins = "pylint_django"
django-settings-module = "nautobot.core.settings"

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "invalid-name",
]
